apartados delete pago

+++ CoreIndustriaHuitzil/Models/Apartados.cs
@@ -10,11 +10,8 @@ namespace CoreIndustriaHuitzil.Models
     {
         public int IdApartado { get; set; }
         public int? IdCliente { get; set; }
-        public int? total { get; set; } 
-        public int? resto { get; set; }
-        //public int? idArticulo { get; set; }
-        //public int? IdTalla { get; set; }
-        //public int? idParent { get; set; }
+        public decimal? total { get; set; } 
+        public decimal? resto { get; set; }
         public DateTime Fecha { get; set; }
         public DateTime? FechaEntrega { get; set; }
         public string? Status { get; set; }
